requirements.txt
gradio>=4.0.0
yt-dlp
opencv-python
mediapipe
ultralytics
faster-whisper
pyannote.audio
ollama
piper-tts
TTS
pydub
scipy
numpy
Pillow
moviepy
ffmpeg-python
sort-tracker-py
langchain
langchain-community
